+++ PREFS/Test/test_1.py
@@ -7,37 +7,41 @@ sys.path.append(os.path.abspath(os.path.join('..')))
 import __init__ as PREFS
 
 ### Test single preferences and without cascade
-prefs = lambda: PREFS.ReadPREFSFile("prefs1")
+prefs = lambda: PREFS.read_prefs_file("prefs1")
 UserPrefs = PREFS.PREFS(prefs, filename="Prefs/prefs1", dictionary=True, interpret=True, verbose=False, cascade=True)#, filterPrefs=filterPrefs) # Change (dictionary, interpret, debug) to True to test it.
 
 def test_reading_overwrite():
-	UserPrefs.OverWritePrefs()
+	UserPrefs.overwrite_prefs()
 	
-	assert UserPrefs.ReadPrefs() == prefs() # Test ReadPrefs() function
+	assert UserPrefs.read_prefs() == prefs() # Test ReadPrefs() function
 	assert UserPrefs.file == prefs() #Test file attribute
 
 def test_writeprefs():
-	UserPrefs.WritePrefs("theme", "dark")
+	UserPrefs.write_prefs("theme", "dark")
 
 	assert UserPrefs.file == {"theme": "dark"}
 
 def test_changefilename_deletefile():
-	UserPrefs.ChangeFilename("prefs")
+	UserPrefs.change_filename("prefs")
 	assert os.path.isfile(f"{UserPrefs.filename}.{UserPrefs.extension}") == True
 
-	UserPrefs.DeleteFile()
+	UserPrefs.delete_file()
 	assert os.path.isfile(f"{UserPrefs.filename}.{UserPrefs.extension}") == False
 
-def test_json():
-	UserPrefs.ConvertToJson()
-	data = PREFS.ReadJsonFile("Prefs/prefs1")
+def test_json_yaml():
+	UserPrefs.convert_to_json()
+	data = PREFS.read_json_file(filename="Prefs/prefs1", extension="json")
+
+	UserPrefs.convert_to_yaml()
+	data1 = PREFS.read_yaml_file("Prefs/prefs1")
 
 	assert UserPrefs.file == data
+	assert UserPrefs.file == data1
 
 if __name__ == "__main__":
 	test_reading_overwrite()
 	test_writeprefs()
-	test_json()
+	test_json_yaml()
 
 	test_changefilename_deletefile() # This will delete the file so most times you won't see the file